requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiomultiprocess==0.9.0
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==21.2.0
aws-cdk.assets==1.168.0
aws-cdk.aws-acmpca==1.168.0
aws-cdk.aws-apigateway==1.168.0
aws-cdk.aws-applicationautoscaling==1.168.0
aws-cdk.aws-autoscaling==1.168.0
aws-cdk.aws-autoscaling-common==1.168.0
aws-cdk.aws-autoscaling-hooktargets==1.168.0
aws-cdk.aws-certificatemanager==1.168.0
aws-cdk.aws-cloudformation==1.168.0
aws-cdk.aws-cloudfront==1.168.0
aws-cdk.aws-cloudwatch==1.168.0
aws-cdk.aws-cloudwatch-actions==1.168.0
aws-cdk.aws-codebuild==1.168.0
aws-cdk.aws-codecommit==1.168.0
aws-cdk.aws-codeguruprofiler==1.168.0
aws-cdk.aws-codepipeline==1.168.0
aws-cdk.aws-codestarnotifications==1.168.0
aws-cdk.aws-cognito==1.168.0
aws-cdk.aws-dynamodb==1.168.0
aws-cdk.aws-ec2==1.168.0
aws-cdk.aws-ecr==1.168.0
aws-cdk.aws-ecr-assets==1.168.0
aws-cdk.aws-ecs==1.168.0
aws-cdk.aws-efs==1.168.0
aws-cdk.aws-eks==1.168.0
aws-cdk.aws-elasticloadbalancing==1.168.0
aws-cdk.aws-elasticloadbalancingv2==1.168.0
aws-cdk.aws-elasticloadbalancingv2-targets==1.168.0
aws-cdk.aws-events==1.168.0
aws-cdk.aws-events-targets==1.168.0
aws-cdk.aws-globalaccelerator==1.168.0
aws-cdk.aws-iam==1.168.0
aws-cdk.aws-imagebuilder==1.168.0
aws-cdk.aws-kinesis==1.168.0
aws-cdk.aws-kinesisfirehose==1.168.0
aws-cdk.aws-kms==1.168.0
aws-cdk.aws-lambda==1.168.0
aws-cdk.aws-logs==1.168.0
aws-cdk.aws-route53==1.168.0
aws-cdk.aws-route53-targets==1.168.0
aws-cdk.aws-s3==1.168.0
aws-cdk.aws-s3-assets==1.168.0
aws-cdk.aws-sam==1.168.0
aws-cdk.aws-secretsmanager==1.168.0
aws-cdk.aws-servicediscovery==1.168.0
aws-cdk.aws-signer==1.168.0
aws-cdk.aws-sns==1.168.0
aws-cdk.aws-sns-subscriptions==1.168.0
aws-cdk.aws-sqs==1.168.0
aws-cdk.aws-ssm==1.168.0
aws-cdk.aws-stepfunctions==1.168.0
aws-cdk.aws-stepfunctions-tasks==1.168.0
aws-cdk.aws-wafv2==1.168.0
aws-cdk.cloud-assembly-schema==1.168.0
aws-cdk.core==1.168.0
aws-cdk.custom-resources==1.168.0
aws-cdk.cx-api==1.168.0
aws-cdk.lambda-layer-awscli==1.168.0
aws-cdk.lambda-layer-kubectl==1.168.0
aws-cdk.lambda-layer-node-proxy-agent==1.168.0
aws-cdk.region-info==1.168.0
bc-python-hcl2==0.3.39
beautifulsoup4==4.11.1
boto3==1.22.4
botocore==1.25.4
cached-property==1.5.2
cachetools==5.0.0
cattrs==1.8.0
cdk-expects-matcher==0.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
checkov==2.0.1102
click==8.1.3
click-option-group==0.5.3
cloudsplaining==0.5.0
colorama==0.4.4
ConfigArgParse==1.5.3
constructs==3.3.161
contextlib2==21.6.0
cyclonedx-python-lib==0.12.3
decorator==5.1.1
deep-merge==0.0.4
detect-secrets==1.2.0
docker==5.0.3
dockerfile-parse==1.2.0
dpath==1.5.0
expects==0.9.0
frozenlist==1.3.0
gitdb==4.0.7
GitPython==3.1.24
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.0
jsii==1.63.2
jsonpath-ng==1.5.3
jsonschema==3.2.0
junit-xml==1.9
lark==1.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8
packageurl-python==0.9.9
packaging==21.0
pluggy==1.0.0
ply==3.11
policy-sentry==0.12.3
policyuniverse==1.5.0.20220426
prettytable==3.2.0
publication==0.0.3
py==1.10.0
pycares==4.1.2
pycep-parser==0.3.4
pycparser==2.21
pyparsing==2.4.7
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
s3transfer==0.5.2
schema==0.7.5
semantic-version==2.9.0
six==1.16.0
smmap==4.0.0
soupsieve==2.3.2.post1
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
tqdm==4.64.0
typeguard==2.13.3
types-setuptools==57.4.14
types-toml==0.10.6
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.2
yarl==1.7.2
zipp==3.8.0
